requirements.txt
aiohttp>=3.5.4
requests>=2.21.0
aiocfscrape>=0.0.7
Flask>=1.0.2
flask_dance>=1.3.0
google_api_python_client>=1.7.7
pymysql>=0.9.3
spacy>=2.0.18
tqdm>=4.29.1
tweepy>=3.7.0
wordcloud>=1.5.0
yake>=0.3.7
python-dotenv
authlib>=0.14.1
six
gunicorn==20.0.4
